+++ inbm-vision/Changelog.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## NEXT - ?
+## 3.0.13 - 2022-08-07
+
+No changes.
 
 ## 3.0.12 - 2022-07-14